+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Master
 
+* Set memory default for Node builds (https://github.com/heroku/heroku-buildpack-ruby/pull/861)
 * Default Ruby version is now 2.5.5, was previously 2.5.3 (https://github.com/heroku/heroku-buildpack-ruby/pull/863)
+
 
 ## v200 (3/7/2019)